tools/requirements.txt
python-hcl2
markdown
pyyaml
diagrams
boto3